Improve profile page main
