+++ frontend/src/javascripts/components/index/user-only.js
@@ -5,7 +5,7 @@ import Header from '../organisms/header'
 import Mypage from '../organisms/mypage-main'
 import ProjectPage from '../organisms/project-page-main'
 import HistoryPage from '../organisms/history-page-main'
-import ProfilePage from '../organisms/profile-page-main'
+import ProfilePageMain from '../organisms/profile-page-main'
 import Footer from '../organisms/footer'
 
 class UserOnly extends Component {
@@ -29,7 +29,11 @@ class UserOnly extends Component {
             path={`${url}/:userId/history`}
             component={HistoryPage}
           />
-          <Route exact path={`${url}/:userId/edit`} component={ProfilePage} />
+          <Route
+            exact
+            path={`${url}/:userId/edit`}
+            component={ProfilePageMain}
+          />
           <Route exact path={`${url}/:userId`} component={Mypage} />
           <Route render={() => <h2>404 Not Found</h2>} />
         </Switch>
