+++ frontend/src/stylesheets/setting_page.scss
@@ -18,10 +18,12 @@
   .avatar-wrapper {
     margin: 0 auto;
     display: flex;
+    flex-direction: column;
+    flex-wrap: wrap;
     justify-content:center;
     align-items: center;
 
-    .avatar-container {
+    .user-img-container {
       margin: 0;
       width: 100px;
       height: 100px;
@@ -29,16 +31,16 @@
       border-radius: 50%;
       padding: 15px 0;
 
-      .avatar {
+      .user-img {
         width: 70px;
         height: 70px;
         border-radius: 50%;
       }
     }
-  }
-  .for-avatar {
-    display: block;
-    margin-bottom: 30px;
-    cursor: pointer;
+    .for-avatar {
+      display: block;
+      margin-bottom: 30px;
+      cursor: pointer;
+    }
   }
 }
